+++ CustomBlocks/DataStructs.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Runtime.InteropServices;
 using UnityEngine;
 
 
@@ -218,5 +219,33 @@ namespace DestinyCustomBlocks
             // Create and return the sprite.
             return Sprite.Create(iconTex, new Rect(0, 0, 512, 512), new Vector2(0.5f, 0.5f));
         }
+    }
+
+    [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
+    public class OpenFileName
+    {
+        public int structSize = 0;
+        public IntPtr dlgOwner = IntPtr.Zero;
+        public IntPtr instance = IntPtr.Zero;
+        public string filter = null;
+        public string customFilter = null;
+        public int maxCustFilter = 0;
+        public int filterIndex = 0;
+        public string file = null;
+        public int maxFile = 0;
+        public string fileTitle = null;
+        public int maxFileTitle = 0;
+        public string initialDir = null;
+        public string title = null;
+        public int flags = 0;
+        public short fileOffset = 0;
+        public short fileExtension = 0;
+        public string defExt = null;
+        public IntPtr custData = IntPtr.Zero;
+        public IntPtr hook = IntPtr.Zero;
+        public string templateName = null;
+        public IntPtr reservedPtr = IntPtr.Zero;
+        public int reservedInt = 0;
+        public int flagsEx = 0;
     }
 }